Restructure project into internal package
Moved API files to internal/websocket directory for better encapsulation and changed import paths accordingly. This enhances the project's structure by segregating internal implementation details.
